requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
beautifulsoup4
pypdf
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
langchain-community
sentence_transformers
langchain_huggingface
faiss-cpu
streamlit
langchain_groq
langchain_core